+++ CHANGELOG.md
@@ -6,6 +6,17 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [1.1.1] - 2023-09-13
+
+### Fixed
+
+- Special characters like `#`, `$`, `%`, etc., now work in character names and are valid.
+- Per specification, character names like `23` do not work but numbers in character names work regardless of order e.g. `R2D2` or `11A`.
+
+### Changed
+
+- You can force a character name with all numbers e.g. `123`, if desired.
+
 ## [1.1.0] - 2023-06-26
 
 ### Added